ci(codeql): move hsh-kms inline tests to tests/ + Noyalib-Standard README
CodeQL's `rust/hard-coded-cryptographic-value` heuristic kept
flagging the deterministic byte-literal fixtures that the inline
`#[cfg(test)] mod tests` in `crates/hsh-kms/src/lib.rs` passed
into `Pepper::apply`. Test-fixture byte literals are categorically
not security findings; the `.github/codeql/codeql-config.yml`
path-exclusion correctly skips `crates/*/tests/**` but cannot
reach into a `mod tests` inside `src/`. Move those three tests
to `tests/coverage.rs` instead.

Also refactor README.md to the Noyalib documentation standard:
- TOC switched from comma-joined inline links to the bulleted
  'link — one-line description' shape, grouped into the four
  Noyalib phases (Getting started / Library reference / Operational).
- 8-row installation channel matrix (Cargo / source / Homebrew /
  Arch AUR / Scoop / Debian / GHCR / shell installer).
- New 'Why this approach?' section consolidating the five
  architectural choices (auto-rehash, zero-unsafe, peppered HMAC,
  FIPS fail-closed, constant-time) with ADR cross-links.
- New 'One-minute migration' section with `diff`-style before/after
  blocks for argonautica, rust-argon2, and bcrypt.
- New per-context quick-links table.
- Security section restructured to RCE prevention → resource
  budgets → defence in depth → supply chain → notes.
- Examples index for all 9 example files in the workspace.
- Quick Start rewritten as a self-contained main() (the previous
  version used rustdoc `#` hidden-line markers that render
  literally on GitHub).
- Benchmark table moved to placeholder cells pending host-pinned
  numbers from cargo bench.


---
THE ARCHITECT ᛫ Sebastien Rousseau ᛫ https://sebastienrousseau.com
THE ENGINE ᛞ EUXIS ᛫ Enterprise Unified Execution Intelligence System ᛫ https://euxis.co
